pyproject.toml
[project]
name = "degree-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.10",
]

[tool.pyright]
extraPaths = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
